backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
anthropic
semgrep
python-jose[cryptography]
pydantic
httpx